+++ src/ytv-feed-fetch-strategy.h
@@ -42,7 +42,7 @@ G_BEGIN_DECLS
  * @st: a #YtvFeedFetchStrategy that caused the callback
  * @mimetype: (null-ok): the response's MIME type
  * @response: (null-ok): the response data
- * @length: length of the response data
+ * @length: length of the response data buffer or -1
  * @err: (null-ok): if an error occurred
  *
  * A callback for when a response is retrieved. The mimetype and the
@@ -51,7 +51,7 @@ G_BEGIN_DECLS
 typedef void (*YtvGetResponseCallback) (YtvFeedFetchStrategy* st,
                                         const gchar* mimetype,
                                         const gint8* response,
-                                        gint64 length,
+                                        gssize length,
                                         GError *err);
 
 struct _YtvFeedFetchStrategyIface
